#4038: Cleaned up automatic parallelization pass into a stablehlo pipeline pass with presharded or unsharded annotations.
