+++ lib/Dialect/StableHLO/Transforms/ShardyCCLToStableHLOCCLPatterns.cpp
@@ -64,7 +64,6 @@ generateDefaultReplicaGroups(const llvm::SmallVector<int64_t> &meshShape) {
 // default replica groups. For example, if the mesh shape = [2, 4], we generate
 // a default replica groups as: [[0,1,2,3],[4,5,6,7]]. Then we transpose it
 // [[0,4],[1,5],[2,6],[3,7]].
-
 static llvm::SmallVector<llvm::SmallVector<int64_t>> transposeReplicaGroups(
     const llvm::SmallVector<llvm::SmallVector<int64_t>> &replicaGroups) {
   assert(!replicaGroups.empty() &&
