+++ lib/RegisterAll.cpp
@@ -115,7 +115,7 @@ void mlir::tt::registerAllPasses() {
   mlir::sdy::registerAllSdyRoundTripImportPassesAndPipeline();
 
   // Register automatic sharding pipeline.
-  mlir::tt::stablehlo::registerAutomaticShardingPipeline();
+  mlir::tt::stablehlo::registerStableHLOPipeline();
 #endif
 }
