+++ test/ttmlir/Dialect/StableHLO/shardy_automatic_parallelization/negative/no_arg_analysis.mlir
@@ -1,5 +1,5 @@
 // REQUIRES: stablehlo
-// RUN: not ttmlir-opt --automatic-sharding-pipeline="mesh-shape=2,4" %s 2>&1 | FileCheck %s
+// RUN: not ttmlir-opt --stablehlo-pipeline="mesh-shape=2,4" %s 2>&1 | FileCheck %s
 
 func.func public @abs(%arg0: tensor<32x48x24x32xf32>) -> tensor<32x48x24x32xf32> {
   %0 = stablehlo.abs %arg0 : tensor<32x48x24x32xf32>
